Switch Refactored
I extracted the validation and the altercation of the validateMove
method. The redundant validation check was placed into the
validateBoardAfterMove method check. This allows one method to perform a
single task and is easier to conceptualize versus the switch statement –
using the Cohesion principle. alterBoardState simply changes the current
boardState based on the validation. You will notice the use of the
Direct Expressions principle with the if/then/else statements and the
order in which they are arranged.